+++ tests/net/dhcpv6/src/main.c
@@ -178,7 +178,7 @@ fail:
 	return NULL;
 }
 
-static void evt_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void evt_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 			struct net_if *iface)
 {
 	ARG_UNUSED(cb);
